% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUtil.R
\name{remove_ANSI}
\alias{remove_ANSI}
\title{Remove ANSI color code}
\usage{
remove_ANSI(strings)
}
\arguments{
\item{strings}{strings, a character vector}
}
\value{
strings with out ANSI characters
}
\description{
Remove ANSI pre-/suffix-fix in a character string.
}
\examples{
remove_ANSI("\033[34m\033[1ma\033[22m\033[39m")
remove_ANSI(c("\033[34m\033[1ma\033[22m\033[39m",
              "\033[34m\033[1mb\033[22m\033[39m"))
}
