% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{checkNameSpace}
\alias{checkNameSpace}
\title{check namespace}
\usage{
checkNameSpace(packages, quietly = FALSE, from = "CRAN")
}
\arguments{
\item{packages}{vector of strings}

\item{quietly}{bool, give you warning on fail?}

\item{from}{string, where this package is from like, "CRAN", "GitHub", only
for output message display purpose}
}
\value{
vector strings, of missing package names
}
\description{
Help you to check if you have certain packages and
return missing package names
}
\examples{
checkNameSpace("ggplot2")
checkNameSpace("random_pkg")
checkNameSpace("random_pkg", quietly = TRUE)
}
