% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{msg}
\alias{msg}
\title{pretty logging message}
\usage{
msg(
  msg,
  level = "INFO",
  .other_color = "white",
  info_text = "INFO",
  warning_text = "WARNING",
  error_text = "ERROR",
  use_color = TRUE
)
}
\arguments{
\item{msg}{a character string of message or a vector of character strings,
each item in the vector presents one line of words}

\item{level}{typically, one of "INFO", "WARNING", "ERROR", not case sensitive.
Other custom levels will work too.}

\item{.other_color}{hex color code or named colors, when levels are not in
"INFO", "WARNING", "ERROR", this value will be used}

\item{info_text}{info level text prefix, use with "INFO" level}

\item{warning_text}{warning level text prefix, use with "WARNING" level}

\item{error_text}{error level text prefix, use with "ERROR" level}

\item{use_color}{bool, default \code{TRUE}, to use color if supported?}
}
\value{
see description
}
\description{
If
\enumerate{
\item \code{use_color = TRUE} or
\item under SPS main package \code{use_crayon}option is \code{TRUE}
\item In a console that supports colors
Then the message will be colorful, other wise no color
}

"INFO" level spawns \code{message}, "WARNING" is \code{warning}, "ERROR" spawns \code{stop},
other levels use \code{cat}.
}
\examples{
msg("this is info")
msg("this is warning", "warning")
try(msg("this is error", "error"))
msg("this is another level", "my level", "green")

}
