% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUtil.R
\name{quiet}
\alias{quiet}
\title{Suppress cat print output}
\usage{
quiet(x)
}
\arguments{
\item{x}{function or expression or value assignment expression}
}
\value{
If your original functions has a return, it will return in
\code{invisible(x)}
}
\description{
Useful if you want to suppress cat or print
}
\examples{
quiet(print(1))
quiet(cat(1))
}
