% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUtil.R
\name{remove_ANSI}
\alias{remove_ANSI}
\title{Remove ANSI color code}
\usage{
remove_ANSI(string)
}
\arguments{
\item{string}{string, length 1}
}
\value{
string with out ANSI characters
}
\description{
Remove ANSI pre-/suffix-fix in a character string.
}
\examples{
remove_ANSI("\033[34m\033[1ma\033[22m\033[39m")
}
